🧹 Make commit template optional to allow for prompt
