+++ src/domain/commands/mod.rs
@@ -1,5 +1,7 @@
 pub mod actions;
 mod actor;
+mod args;
 
 pub use actions::Actions;
 pub use actor::Actor;
+pub use args::CommitArgs;
